student management system completed backend

+++ app/Http/Controllers/StudentController.php
@@ -10,10 +10,14 @@ use App\Http\Resources\ClassResource;
 use App\Http\Resources\StudentResource;
 use App\Http\Requests\StoreStudentRequest;
 use App\Http\Requests\UpdateStudentRequest;
+use Illuminate\Support\Facades\Gate;
 
 class StudentController extends Controller
 {
     public function index(Request $request){
+
+        Gate::authorize('student_access');
+
         $students = Student::search($request)->paginate(10);
         $classes = ClassResource::collection(Classes::all());
         return Inertia::render('Student/Index', [
@@ -25,6 +29,9 @@ class StudentController extends Controller
     }
 
     public function create(){
+
+        Gate::authorize('student_create');
+
         $classes = ClassResource::collection(Classes::all());
         return Inertia::render('Student/Create', [
             'classes' => $classes
@@ -32,6 +39,9 @@ class StudentController extends Controller
     }
 
     public function store(StoreStudentRequest $request){
+
+        Gate::authorize('student_create');
+
         Student::create($request->validated());
         return redirect()->route('students.index');
     }
@@ -45,11 +55,17 @@ class StudentController extends Controller
     }
 
     public function update(UpdateStudentRequest $request, Student $student){
+
+        Gate::authorize('student_edit');
+
         $student->update($request->validated());
         return redirect()->route('students.index');
     }
 
     public function destroy(Student $student){
+
+        Gate::authorize('student_delete');
+
         $student->delete();
         return redirect()->route('students.index');
     }
